Added my first two beginner java projects for you
This is a console-based calculator and a console-based random fortune cookie program. The code is fully commented and I try to comment on each and every line for these beginner projects. I hope to record my journey and also make some good teaching exercises for java beginners. If you are trying to find some exercises then first Read what the program does, try to do it on your own, and if you need help check out how I did it.